Split model into sevaral modules for easier maangement
